Replace SignerFactory interface with extensible struct
Rather than providing alternative implementations of SignerFactory in
order to support other types of keys, this SignerFactory struct can be
extended by registering additional key protobuf handlers. The need for
an interface no longer exists, and the MockSignerFactory is also
unnecessary, because test handlers can be used instead.

This approach makes it easier to select a combination of supported key
types. Key types that are not required can be excluded from the build
through the use of build tags. For example, this change results in
PKCS#11 key support being excluded from the build unless the build tag
"pkcs11" is used, e.g. `go build -t pkcs11`. A key benefit is that this
no longer forces users to install dependencies for all key handlers - only
those that they include in their build.